Revert "server.js: Open local `from-url` URL in separate script"
This reverts commit 91b485c464eff46d7191aaa93dc0852b8107f83c.
