+++ .eslintrc.js
@@ -141,7 +141,6 @@ module.exports = {
         ],
         extensions: ['.js', '.jpg'],
       },
-      node: true,
     },
   },
   globals: {
